Revert "Make -a more efficient. +Now it generates images in reversed order."
This reverts commit a0269a03f220678986f5bd9aef2694fbfc16d20e.

+++ src/Frame.cpp
@@ -13,8 +13,8 @@ namespace am{
         check=get_str(offset,0x4);
         advance(offset,0x4);
 
-        position_x=(short)get_int(offset,0x2);
-        position_y=(short)get_int(offset,0x2);
+        position_x=get_int(offset,0x2);
+        position_y=get_int(offset,0x2);
 
         advance(offset,0x4);
